# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArogyaLive API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
...
  /beds:
    get:
      operationId: getBeds
      tags: [beds]
      summary: Get bed availability for a facility
      parameters:
        - name: facilityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bed records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BedRecord"
  /beds/{id}:
    patch:
      operationId: updateBed
      tags: [beds]
      summary: Update bed occupancy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BedUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BedRecord"
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get alerts
      parameters:
        - name: facilityId
          in: query
          schema:
            type: integer
        - name: since
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /forecast/{facilityId}:
    get:
      operationId: getForecast
      tags: [ai]
      summary: Get AI-powered 7-day stock forecast for a facility
      parameters:
        - name: facilityId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Forecast data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastResult"
  /assistant/query:
    post:
      operationId: queryAssistant
      tags: [ai]
      summary: Query the AI chat assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantQueryInput"
      responses:
        "200":
          description: Answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantAnswer"
  /voice/parse:
    post:
      operationId: parseVoiceCommand
      tags: [ai]
      summary: Parse a voice command into structured action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoiceCommandInput"
      responses:
        "200":
          description: Parsed command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VoiceCommandResult"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, name, username, role]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [facility_staff, district_admin]
        facilityId:
          type: ["integer", "null"]
        facilityName:
          type: ["string", "null"]
        languagePref:
          type: string
          enum: [en, hi, bn]
...
    BedRecord:
      type: object
      required: [id, facilityId, wardName, totalBeds, occupiedBeds]
      properties:
        id:
          type: integer
        facilityId:
          type: integer
        wardName:
          type: string
        totalBeds:
          type: integer
        occupiedBeds:
          type: integer
        availableBeds:
          type: integer
        lastUpdatedAt:
          type: string
    BedUpdate:
      type: object
      required: [occupiedBeds]
      properties:
        occupiedBeds:
          type: integer
...
    Alert:
      type: object
      required: [id, type, severity, message, isRead, createdAt]
      properties:
        id:
          type: integer
        facilityId:
          type: ["integer", "null"]
        facilityName:
          type: ["string", "null"]
        type:
          type: string
        severity:
          type: string
          enum: [info, warning, critical]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    ForecastMedicine:
      type: object
      required: [medicineId, medicineName, currentStock, daysRemaining, forecastData]
      properties:
        medicineId:
          type: integer
        medicineName:
          type: string
        currentStock:
          type: number
        daysRemaining:
          type: ["number", "null"]
        forecastData:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              projected:
                type: number
        aiInsight:
          type: ["string", "null"]
        riskLevel:
          type: string
          enum: [ok, watch, critical]
    ForecastResult:
      type: object
      required: [facilityId, facilityName, medicines]
      properties:
        facilityId:
          type: integer
        facilityName:
          type: string
        medicines:
          type: array
          items:
            $ref: "#/components/schemas/ForecastMedicine"
    AssistantQueryInput:
      type: object
      required: [question]
      properties:
        question:
          type: string
    AssistantAnswer:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        sources:
          type: array
          items:
            type: string
    VoiceCommandInput:
      type: object
      required: [transcript, facilityId]
      properties:
        transcript:
          type: string
        facilityId:
          type: integer
        language:
          type: string
    VoiceCommandResult:
      type: object
      required: [action, confidence]
      properties:
        action:
          type: string
          enum: [add_stock, log_consumption, log_footfall, update_beds, mark_attendance, unknown]
        medicineName:
          type: ["string", "null"]
        medicineId:
          type: ["integer", "null"]
        quantity:
          type: ["number", "null"]
        department:
          type: ["string", "null"]
        count:
          type: ["integer", "null"]
        confidence:
          type: number
        rawTranscript:
          type: string
...